+++ src/sql/__init__.py
@@ -1,4 +1,6 @@
 from .magic import RenderMagic, SqlMagic, load_ipython_extension
+from .error_message import SYNTAX_ERROR
+from .connection import PLOOMBER_DOCS_LINK_STR
 
 __version__ = "0.7.5dev"
 
@@ -7,4 +9,6 @@ __all__ = [
     "RenderMagic",
     "SqlMagic",
     "load_ipython_extension",
+    "SYNTAX_ERROR",
+    "PLOOMBER_DOCS_LINK_STR",
 ]