+++ src/tests/test_magic.py
@@ -19,6 +19,10 @@ from sql.run import ResultSet
 from sql import magic
 
 from conftest import runsql
+from sql.connection import PLOOMBER_DOCS_LINK_STR
+from ploomber_core.exceptions import COMMUNITY
+
+COMMUNITY = COMMUNITY.strip()
 
 
 def test_memory_db(ip):
@@ -722,7 +726,7 @@ def test_autolimit(ip):
     assert len(result) == 1
 
 
-invalid_connection_string = """
+invalid_connection_string = f"""
 No active connection.
 
 To fix it:
@@ -734,8 +738,8 @@ OR
 
 Set the environment variable $DATABASE_URL
 
-For technical support: https://ploomber.io/community
-Documentation: https://jupysql.ploomber.io/en/latest/connecting.html
+{PLOOMBER_DOCS_LINK_STR}
+{COMMUNITY}
 """
 
 
@@ -746,14 +750,14 @@ def test_error_on_invalid_connection_string(ip_empty, clean_conns):
     assert isinstance(result.error_in_exec, UsageError)
 
 
-invalid_connection_string_format = """\
+invalid_connection_string_format = f"""\
 Can't load plugin: sqlalchemy.dialects:something
 
 To fix it, make sure you are using correct driver name:
 Ref: https://docs.sqlalchemy.org/en/20/core/engines.html#database-urls
 
-For technical support: https://ploomber.io/community
-Documentation: https://jupysql.ploomber.io/en/latest/connecting.html
+{PLOOMBER_DOCS_LINK_STR}
+{COMMUNITY}
 """  # noqa
 
 
@@ -764,32 +768,21 @@ def test_error_on_invalid_connection_string_format(ip_empty, clean_conns):
     assert isinstance(result.error_in_exec, UsageError)
 
 
-invalid_connection_string_existing_conns = """
-Can't load plugin: sqlalchemy.dialects:something
-
-To fix it, make sure you are using correct driver name:
-Ref: https://docs.sqlalchemy.org/en/20/core/engines.html#database-urls
-
-For technical support: https://ploomber.io/community
-Documentation: https://jupysql.ploomber.io/en/latest/connecting.html
-"""  # noqa
-
-
 def test_error_on_invalid_connection_string_with_existing_conns(ip_empty, clean_conns):
     ip_empty.run_cell("%sql sqlite://")
     result = ip_empty.run_cell("%sql something://")
 
-    assert invalid_connection_string_existing_conns.strip() == str(result.error_in_exec)
+    assert invalid_connection_string_format.strip() == str(result.error_in_exec)
     assert isinstance(result.error_in_exec, UsageError)
 
 
-invalid_connection_string_with_possible_typo = """
+invalid_connection_string_with_possible_typo = f"""
 Can't load plugin: sqlalchemy.dialects:sqlit
 
 Perhaps you meant to use driver the dialect: "sqlite"
 
-For technical support: https://ploomber.io/community
-Documentation: https://jupysql.ploomber.io/en/latest/connecting.html
+{PLOOMBER_DOCS_LINK_STR}
+{COMMUNITY}
 """  # noqa
 
 
@@ -800,6 +793,29 @@ def test_error_on_invalid_connection_string_with_possible_typo(ip_empty, clean_c
     assert invalid_connection_string_with_possible_typo.strip() == str(
         result.error_in_exec
     )
+    assert isinstance(result.error_in_exec, UsageError)
+
+
+invalid_connection_string_duckdb = f"""
+An error happened while creating the connection: connect(): incompatible function arguments. The following argument types are supported:
+    1. (database: str = ':memory:', read_only: bool = False, config: dict = None) -> duckdb.DuckDBPyConnection
+
+Invoked with: kwargs: host='invalid_db'.
+
+To fix it:
+
+Pass a valid connection string:
+    Example: %sql postgresql://username:password@hostname/dbname
+
+{PLOOMBER_DOCS_LINK_STR}
+{COMMUNITY}
+"""  # noqa
+
+
+def test_error_on_invalid_connection_string_duckdb(ip_empty, clean_conns):
+    result = ip_empty.run_cell("%sql duckdb://invalid_db")
+
+    assert invalid_connection_string_duckdb.strip() == str(result.error_in_exec)
     assert isinstance(result.error_in_exec, UsageError)
 
 
@@ -916,8 +932,8 @@ def test_save_with_non_existing_table(ip, capsys):
 def test_save_with_bad_query_save(ip, capsys):
     ip.run_cell("%sql --save my_query SELECT * non_existing_table")
     ip.run_cell("%sql --with my_query SELECT * FROM my_query")
-    out, _ = capsys.readouterr()
-    assert '(sqlite3.OperationalError) near "non_existing_table": syntax error' in out
+    out, err = capsys.readouterr()
+    assert '(sqlite3.OperationalError) near "non_existing_table": syntax error' in err
 
 
 def test_interact_basic_data_types(ip, capsys):