+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # CHANGELOG
 
 ## 0.7.5dev
-* [Feature] Using native DuckDB `.df()` method when using `autopandas` 
-
+* [Feature] Using native DuckDB `.df()` method when using `autopandas`
 * [Doc] documenting `%sqlcmd tables`/`%sqlcmd columns`
 * [Feature] Better error messages when function used in plotting API unsupported by DB driver (#159)
 * [Fix] Fix the default value of %config SqlMagic.displaylimit to 10 (#462)
+* [Feature] Detailed error messages when syntax error in SQL query, postgres connection password missing or inaccessible, invalid DuckDB connection string (#229)
+
 
 ## 0.7.4 (2023-04-28)
 No changes